feat: added new pages and separated code in more files
