+++ index.js
@@ -4,6 +4,9 @@ const mongoose = require('mongoose');
 const config = require('./config/config')[env];
 const express = require('express');
 const indexRouter = require('./routes');
+const authRouter = require('./routes/auth');
+const cubeRouter = require('./routes/cube');
+const accessoryRouter = require('./routes/accessory');
 const app = express();
 
 mongoose.connect(config.databaseUrl,
@@ -22,6 +25,17 @@ mongoose.connect(config.databaseUrl,
 
 require('./config/express')(app);
 
+app.use('/', authRouter);
+app.use('/', cubeRouter);
+app.use('/', accessoryRouter);
 app.use('/', indexRouter);
+
+
+app.get('*', (req, res) => {
+    res.render('404', {
+        title: 'Not Found | Cube Workshop'
+    });
+});
+
 
 app.listen(config.port, console.log(`Listening on port ${config.port}! Now its up to you...`));
