+++ views/loginPage.hbs
@@ -1,0 +1,13 @@
+<main>
+  <h1>Login Form</h1>
+  <div class="form">
+    <form action='/' method="POST">
+      <label for="username">Username</label>
+      <input type="text" id="username" name="username">
+
+      <label for="password">Password</label>
+      <input type="password" id="password" name="password">
+      <input type="submit" value="Login">
+    </form>
+  </div>
+</main>
